requirements.txt
pinecone-haystack
haystack-ai
python-dotenv
pathlib 
Flask


-e .
